Chore: fixup docutils
